+++ src/openassetio-python/tests/package/test_imports.py
@@ -1,5 +1,5 @@
 #
-#   Copyright 2013-2021 The Foundry Visionmongers Ltd
+#   Copyright 2013-2023 The Foundry Visionmongers Ltd
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -132,9 +132,6 @@ class Test_test_imports:
     def test_importing_manager_succeeds(self):
         from openassetio.test import manager
 
-    def test_importing_specifications_succeeds(self):
-        from openassetio.test import specifications
-
 
 class Test_test_manager_imports:
     def test_importing_apiComplianceSuite_succeeds(self):
@@ -142,6 +139,3 @@ class Test_test_manager_imports:
 
     def test_importing_harness_succeeds(self):
         from openassetio.test.manager import harness
-
-    def test_importing_specifications_succeeds(self):
-        from openassetio.test.manager import specifications